newsrecomendation/DataCollector/requirements.txt
newsapi-python==0.2.7
pymongo==4.8.0
mongoengine==0.28.2
pika==1.3.2